Feat: 대시보드 제목/색상 수정 Hook으로 분리
